+++ terminal.c
@@ -6441,7 +6441,7 @@ int term_get_userpass_input(Terminal *term, prompts_t *p,
 	 */
 	{
 	    int i;
-	    for (i = 0; i < p->n_prompts; i++)
+	    for (i = 0; i < (int)p->n_prompts; i++)
 		memset(p->prompts[i]->result, 0, p->prompts[i]->result_len);
 	}
     }
